Update Main.java
Guys I just found one problem and fixed it - for all the methods in main, there need to include a Hotel object as parameter - in our case, the hotel burger. if you don't include the hotel parameter, the program will lose the hotel object created by the file reader. we are basically passing the hotel burger all along to store data. Please be aware of that in the future while implementing the other menus of the program.